Cluster conflict and discovered servers srvPool issue
When there is not an explicitly set cluster name, the gossiping
based cluster name consensus will cause the `connect_urls` field
to be reset a number of times for the client.

As a result, even though the cluster name has been decided and the
cluster has formed a mesh, the client will have dropped the discovered
servers from its internal srvPool.

Signed-off-by: Waldemar Quevedo <wally@nats.io>