Decouple some tests that depend on server
This starts the move of some tests that use internals
of the client but also depend on running the server in tests
which brings the nats-server as a dependency for the nats package
when running `go test github.com/nats-io/nats.go`.

The package `internals/test` is introduced to use a private interface
that can be shared among tests in the same module but not accessible
by users of the library.

Signed-off-by: Waldemar Quevedo <wally@nats.io>